+++ deploy/job/development/config.yaml
@@ -15,9 +15,11 @@ logger:
   file_max_age_in_days: 7
 
 temporal:
-  greeting_queue_name: "greeting"
   start_to_close_timeout: 10
   initial_interval: 1
   backoff_coefficient: 2
   maximum_interval: 5
   maximum_attempts: 10
+  namespace: "default"
+  temporal_host_port: "localhost:7233"
+
