+++ jobapp/config.go
@@ -1,7 +1,7 @@
 package jobapp
 
 import (
-	"github.com/gocastsian/roham/jobapp/service/job"
+	"github.com/gocastsian/roham/adapter/temporal"
 	httpserver "github.com/gocastsian/roham/pkg/http_server"
 	"github.com/gocastsian/roham/pkg/logger"
 	"time"
@@ -11,5 +11,5 @@ type Config struct {
 	HTTPServer           httpserver.Config `koanf:"http_server"`
 	Logger               logger.Config     `koanf:"logger"`
 	TotalShutdownTimeout time.Duration     `koanf:"total_shutdown_timeout"`
-	Temporal             job.Config        `koanf:"temporal"`
+	Temporal             temporal.Config   `koanf:"temporal"`
 }
