+++ jobapp/service/job/shared.go
@@ -1,0 +1,5 @@
+package job
+
+const (
+	GREETING_QUEUE_NAME = "greeting"
+)
